refactor(metadata): use same host name lookup as in sitemap.xml
