+++ test/unit/html.pre.test.js
@@ -18,7 +18,8 @@ const { pre } = require('../../src/html.pre.js');
 const request = {
   headers: {
     host: 'foo.bar',
-    'x-cdn-url': 'https://foo.bar/baz.html',
+    'x-hlx-pages-host': 'www.foo.bar',
+    'x-forwarded-host': 'www.baz.bar',
   },
   url: '/baz.html',
 };
@@ -139,7 +140,7 @@ describe('Testing pre.js', () => {
     assert.ok(context.content.meta.description.endsWith('...'));
   });
 
-  it('Meta url uses x-cdn-url if available', () => {
+  it('Meta url uses x-hlx-pages-host if available', () => {
     const dom = new JSDOM('<html><head><title>Foo</title></head><body></body></html');
     const context = {
       content: {
@@ -151,15 +152,15 @@ describe('Testing pre.js', () => {
     pre(context);
 
     assert.ok(context.content.meta.url);
-    assert.equal(context.content.meta.url, request.headers['x-cdn-url']);
+    assert.equal(context.content.meta.url, `https://${context.request.headers['x-hlx-pages-host']}${context.request.url}`);
   });
 
-  it('Meta url uses host header and request.url if no x-cdn-url available', () => {
+  it('Meta url uses x-forwarded-host header if no x-hlx-pages-host available', () => {
     const req = {
       ...request,
       headers: {
         ...request.headers,
-        'x-cdn-url': undefined,
+        'x-hlx-pages-host': undefined,
       },
     };
     const dom = new JSDOM('<html><head><title>Foo</title></head><body></body></html');
@@ -172,7 +173,7 @@ describe('Testing pre.js', () => {
     };
     pre(context);
 
-    assert.equal(context.content.meta.url, `https://${context.request.headers.host}${context.request.url}`);
+    assert.equal(context.content.meta.url, `https://${context.request.headers['x-forwarded-host']}${context.request.url}`);
   });
 
   it('Meta imageUrl uses content.image', () => {
